chore: toolchain pin, lint cleanup, dep audit
- add rust-toolchain.toml (stable) + rust-version 1.70 in Cargo.toml
- fix 7 clippy warnings (is_empty, redundant cast, struct-update init),
  MSRV-safe `% 250`
- gate CI on --all-targets --all-features
- bump event-listener 5.4.1 -> 5.4.2 (RUSTSEC-2026-0221)
- add deny.toml license/advisory/source config
- add weekly security.yml (cargo audit + deny)
- no unused deps (cargo machete)

+++ src/bench.rs
@@ -111,7 +111,7 @@ pub fn run_bench() {
     let cpu_burst1 = proc_cpu_self();
     let burst_cpu_ms = cpu_delta_ms(cpu_burst0, cpu_burst1);
     let burst_util = if burst_wall.as_millis() > 0 {
-        (burst_cpu_ms as f64) * 100.0 / (burst_wall.as_millis() as f64)
+        burst_cpu_ms * 100.0 / (burst_wall.as_millis() as f64)
     } else {
         0.0
     };
